requirements.txt
whisper
torch
requests
python-dotenv
